+++ src/service/delivery/index.js
@@ -1,0 +1,22 @@
+const OsduV2DeliveryService = require('./v2');
+
+/**
+ * Class that provides named access to OSDU Delivery endpoints
+ * - [Service Documentation]{@link https://community.opengroup.org/osdu/documentation/-/wikis/OSDU-(C)/Design-and-Implementation/Ingestion-and-Enrichment-Detail/R2-Delivery}
+ * - [API Documentation]{@link https://community.opengroup.org/osdu/platform/system/delivery/-/blob/master/docs/api/delivery.yaml}
+ * @class
+ * @category Services
+ * @subcategory Delivery
+ */
+class OsduDeliveryService {
+    /**
+     * @constructor
+     * @param {BaseOsduClient} osdu_client - An implementation of the OSDU client class to broker communication with the OSDU API
+     * @param {string} data_partition - The data partition against which requests will be made
+     */
+    constructor(osdu_client, data_partition) {
+        this.V2 = new OsduV2DeliveryService(osdu_client, data_partition);
+    }
+}
+
+module.exports = OsduDeliveryService;