+++ src/service/entitlements/index.js
@@ -1,25 +1,20 @@
-const OsduR2BaseService = require('./base');
-
 /**
- * Class that provides named access to OSDU R2 Entitlements endpoints
+ * Class that provides named access to OSDU Entitlements endpoints
  * - Not yet supported
  * - [Service Documentation]{@link https://community.opengroup.org/osdu/documentation/-/wikis/Releases/R2.0/Services/Entitlements/Tutorial}
  * - [API Documentation]{@link https://community.opengroup.org/osdu/documentation/-/wikis/Releases/R2.0/Services/Entitlements/Swagger}
  * @class
  * @category Services
- * @subcategory R2
+ * @subcategory Entitlements
  */
-class OsduR2EntitlementsService extends OsduR2BaseService {
+class OsduEntitlementsService {
     /**
      * @constructor
      * @param {BaseOsduClient} osdu_client - An implementation of the OSDU client class to broker communication with the OSDU API
      * @param {string} data_partition - The data partition against which requests will be made
      */
     constructor(osdu_client, data_partition) {
-        super(osdu_client, data_partition);
     }
-    
-    // TODO: Implement Entitlements
 }
 
-module.exports = OsduR2EntitlementsService;
+module.exports = OsduEntitlementsService;