+++ types/models/index.d.ts
@@ -1,4 +1,5 @@
 declare const _exports: {
+    OsduSchemaQueryBuilder: typeof import("./schema/osduSchemaQueryBuilder");
     OsduQueryBuilder: typeof import("./query/osduQueryBuilder");
     OsduQueryExpression: typeof import("./query/osduQueryExpression");
 };